sql/planner/optimizer/cost/mod.rs -> sql/planner/optimizer/cost/cost.rs
